requirements.txt
PyYAML==5.3.1
python-docx==0.8.10
requests==2.24.0
rocketchat-API==1.9.1
urllib3==1.25.10